backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pandas==2.0.3
pillow==10.1.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.17.0
pydantic==2.5.0
aiofiles==23.2.1
